backend/requirements.txt
boto3
python-dotenv
faiss-cpu
langgraph
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.2
langchain-google-genai==1.0.10